Modal restructure/clean-up
