+++ src/App.css
@@ -51,7 +51,7 @@ h2 {
   align-items: center;
   justify-content: center;
   background: rgba(77, 77, 77, .7);
-  transition: 1s;
+  transition: all 1s;
 }
 
 .modal:target {
